Ensure disposables are being tracked by the entities that own them (#3142)
### Motivation

While studying #3137, I noticed there were some disposables not being handled properly. Whatever we push to `context.subscriptions` will only be disposed when the extension is deactivated, which means we continue to waste memory and resources unnecessarily.

We need the client and workspace to better own their disposables.

### Implementation

I tried minimizing the times where we invoke `context.subscriptions`, so that problems of this nature are easier to spot.

There were two command registrations that were leaking and two client hooks.