Allow server to produce data telemetry (#3150)
### Motivation

Allow the server to produce data telemetry, in addition to error telemetry. This is useful for both the server and add-ons to track adoption of certain features or how they are used.

Here's an example of how it would be used.
```ruby
Notification.telemetry(type: "data", eventName: "tapioca.dsl", data: { type: "counter" })
```

### Implementation

Just added a branch for the data telemetry. Since the server and add-ons can send arbitrary data, I added runtime validation to prevent sending bad data to telemetry.

### Automated Tests

It's unfortunately not possible to test this because the client doesn't expose a way to fake an incoming notification. I did test it manually though.