Pass machine ID to server for telemetry (#3157)
### Motivation

Pass the unique machine identifier to the server as an initialization option, so that it can be used to estimate number of unique users on server telemetry events.

Reminder that we [do not collect any telemetry by default](https://github.com/Shopify/ruby-lsp/tree/main/vscode#telemetry), unless a VS Code extension implements the `getTelemetrySenderObject` command.

### Implementation

Started passing the value from the extension and saving it in the global state.

### Automated Tests

Added a test.