Avoid indexing client managed documents twice (#3159)
### Motivation

Noticed this bug when investigating some weirdness in the Tapioca add-on, which I accidentally introduced when we started indexing files before they are saved.

Once the file is saved and we receive a did change notification, we should not try to index files that are being managed by the client (open documents) because those are indexed already at that point.

Additionally, if we happen to receive a late `CREATED` notification after the client already informed the server that the document is opened, then we need to `handle_change` instead of `index_single` to avoid duplicate entries while still adjusting the require path tree.

### Implementation

Started checking if the document was already claimed by the client before doing work on the did changed notification.

### Automated Tests

Added two tests that will prevent regression.