Accept module/class node's constant path for constant name (#3183)
### Motivation

It's useful to pass a module/class `constant_path` directly into `constant_name`, but that means our typing isn't satisfied because `constant_path` can be a missing node or a call node too.

### Implementation

Started accounting for those possible cases so that we can pass a constant path directly.

### Automated Tests

Added tests.