Never include workspace name as part of relative test URI (#3181)
### Motivation

I noticed that the new test discovery was not working for the Ruby LSP's own workspace config. The reason was because `vscode.workspace.asRelativePath` defaults to returning the workspace name as part of the path, which was generating an invalid test file path.

### Implementation

We just need to pass `false` as the second argument to never include the workspace name when getting a relative path and then everything works.

### Automated Tests

We need to stub `asRelativePath` in the tests to force it to return our fake workspace folders, which is why the test didn't catch this in the first place.

I'm not sure how we could restructure the test to still have the fake folder and avoid stubbing the `asRelativePath` function. For now, let's just ensure the bug is fixed.