Add test items collection builder (#3169)
### Motivation

To be able to provide complete test explorer support, we should create a collection builder with a corresponding interface object that allow us to provide VS Code with all of the information needed to provide full functionality for tests.

This PR adds an initial version of a test item response builder that we can enhance as needed to represent test items.

### Implementation

Created a test item object that matches what VS Code expects in the explorer. Also created a response builder to along with it.

### Automated Tests

Added some tests.