Resolve test items with server request (#3186)
### Motivation

Closes #3173

Start firing our new custom request to discover which tests are available in a given file.

### Implementation

We fire the request and then recursively add test items to their respective collections. If a document is saved, we check to see if we created a test item for it and then we clear its children (the tests defined inside of the file) and re-trigger the request.

I chose to use on save because on change would put a lot of pressure on the server.

### Automated Tests

Added tests.