Resolve test items with server request (#3186)
* Add discover tests custom request (#3180)

### Motivation

Closes #3171

Add a new custom request to discover tests in a specific document. This request will instantiate all listeners and collect all of the discovered groups and examples as test items for the editor.

### Implementation

- Created the new request
- Implement a listener dedicated to the test style syntax only (spec will be a separate listener)
- Ensured to use ancestor linearization to determine the test framework

### Automated Tests

Added tests.

* Resolve test items with server request (#3186)

### Motivation

Closes #3173

Start firing our new custom request to discover which tests are available in a given file.

### Implementation

We fire the request and then recursively add test items to their respective collections. If a document is saved, we check to see if we created a test item for it and then we clear its children (the tests defined inside of the file) and re-trigger the request.

I chose to use on save because on change would put a lot of pressure on the server.

### Automated Tests

Added tests.

---------

Co-authored-by: vinistock <18742907+vinistock@users.noreply.github.com>