Avoid focusing on test items on document switch (#3197)
### Motivation

VS Code doesn't provide an API for us to know which test item is currently selected in the UI.

This means that, if you have a specific test example selected, it gets unselected when you switch documents due to the callback we added.

### Implementation

It's unfortunate, but I don't think we're able to auto-select the items correctly based on which document is opened.