Add command to show LSP's internal state for diagnosing issues (#3195)
Add command to show LSP's internal state for diagnosing issues (#3195)

### Motivation

Add a command to trigger the new custom diagnose request, so that it's easy to see the information at a glance.

### Implementation

It's the same implementation as show syntax tree, we use a custom resource to show a read-only document containing the relevant information.

### Manual Tests

1. Launch the extension on this branch
2. On the second VS Code window, open the Ruby LSP code itself (you need the server changes to verify all pieces together)
3. Invoke the new command
4. Verify you see the internal state of the server

Co-authored-by: vinistock <18742907+vinistock@users.noreply.github.com>