Fix incorrect RBS highlighting in Ruby code (#3221)
The `#:` grammar rule should only be applied to comments, which means
the capture group should only match if the `#:` is the first non-whitespace
character in the line.

Otherwise, the grammar will incorrectly apply to `#:` in Ruby code as well.

This commit fixes the grammar rule so that it only applies to comments.