Ensure we index non test files under test directories (#3188)
### Motivation

I noticed something with #3180. We can only accurately linearize ancestors for test files if we index files under the test directories that aren't tests, but support implementations.

For example, in our own codebase, we have `lib/ruby_indexer/test/test_case.rb`, which is not a test, but implements the parent class for all tests in the indexer. If we never index that parent class, it's impossible to linearize the ancestors of any of the tests and we can't discover them properly.

### Implementation

I propose we start indexing non test files inside the test directory. In addition to allowing our improved test functionality to work, users often refer to these support declarations when writing the tests themselves.

I also simplified the code, which was hard to follow, while still trying to maintain the original performance gains applied by recent changes.

One small thing is that I couldn't think of a decent way to not exclude `test_case` and `test_helper` from indexing, since both of them are common names that may declare test parent classes and match our file exclusion glob. I created a special case to handle them.

I benchmarked this and it makes no substantial difference in indexing time in Core.

### Automated Tests

Added a test verifying the behaviour changes.