Try to relaunch if gems are modified between install and setup (#3231)
### Motivation

If a user manages to change their Gemfile or uninstall a gem from an external command  while we're in the middle of running bundle install, Bundler may have already instantiated its internal state of the previous version of the Gemfile/dependency states.

This means composing the bundle will succeed, but then when we reach `Bundler.setup`, it will fail because now the Gemfile is in a different state or expected gems are no longer installed.

### Implementation

I propose we perform a single retry to begin the process from scratch if we notice that this happened. That way, the bundle is recomposed and the user will hopefully not change the Gemfile in the middle of launch twice.

### Automated Tests

Added a test.