Consider server test item tags for test explorer (#3235)
### Motivation

Step towards #3175

We will need to ensure that the extension is passing all server tags back when trying to resolve test commands. Additionally, it's useful for the listeners to know if a test item refers to an entire file, since composing the command for those is significantly simpler.

### Implementation

Started including server tags as part of the ones we already assign.

### Automated Tests

Added a new test.