Add basic resolve test command request (#3234)
### Motivation

Step for #3175

This PR adds the basic structure of a new custom request to determine what shell command will execute a group of selected tests in the explorer.

### Implementation

For now, I added only the basic structure to get feedback on the approach. My idea is to allow listeners to tag the items they contribute to discovery, so that they avoid returning anything for the command if the test item doesn't "belong" to them.

Taking add-ons into consideration is a future issue, but I'm trying to make sure we're accounting for what they need.

### Automated Tests

Added a placeholder test.