Run `vscode/activation.rb` script in CI
Given the purpose of this script is to collect information about the
user's Ruby version, we don't know their version yet. Therefore, we must
write the most backwards compatible Ruby we can, or else they may get
errors when the editor's environment includes an older Ruby version than
we expect (e.g. an old system Ruby, in a non-Ruby project).

Running the script in Ci ensures we are compatible at runtime, not just
as a matter of syntax compatibility.