Resolve Minitest commands for test items (#3249)
### Motivation

This PR completes the Minitest portion of #3175. Only Test Unit left to go.

We need to receive a hierarchy of test items from the extension and determine the minimum set of shell commands that can execute those tests.

For Minitest, that means creating some regexes that will match all of the desired groups or examples. Additionally, we need to take test directories and entire files into account as well.

### Implementation

The implementation is a bit complex, but I would like to see what it looks like after implementing Test Unit since there may be opportunities for simplification.

The idea is to first aggregate test items by file path -> test group -> test examples. If a test file has no groups inside, then we're executing the entire file. If the test group has no examples inside, then we're executing the entire group.

We generate a single command for running all test files and directories that don't require specific groups or examples. Those can be executed in a single command.

Then for groups and examples, we need one command per file so that we can created a `--name` regex that will match the selected groups and examples.

### Automated Tests

Added a bunch of tests for different test filtering scenarios.