Tag Test Unit and Minitest spec test items (#3268)
### Motivation

We need to ensure that all test items are properly tagged with the framework they are related to so that we can resolve commands.

We were talking about refactoring this to be a separate attribute in test items, but let's first ensure that all items tag and then we move ahead with the refactor.

### Implementation

Ensured `Minitest::Spec` and `Test::Unit::TestCase` tests are all tagged.

### Automated Tests

Added tests verifying that all test groups and their children are tagged with a framework.