Exclude certain Bundler settings when composing bundle (#3277)
### Motivation

Closes #2440, closes #3196

Certain Bundler settings are better used only when the user is in control, like deciding when to package gems or if Bundler should automatically install binstubs for all gems.

This PR starts ignoring certain settings from the composed environment.

### Implementation

Added a list of ignored settings that we reject before building the environment that we'll use to run bundle install/update.

### Automated Tests

Added tests for both bundle bin and bundle package related settings.