Add new run handler for test controller (#3251)
### Motivation

Closes #3177

This PR adds a new run handler for the test controller, which asks the server for the resolve commands and then executes them with streaming updates.

**Important**: after this PR, there are still some details to iron out and this doesn't fully work out of the box. For example, we need the Test Unit reporter to automatically select the custom one.

We also may need a way for add-ons to register their reporter paths so that they are required automatically.

However, this implements the bulk of the integration, so it will be easier to deal with the other details once this is shipped.

### Implementation

1. Added the JSON RPC package as an explicit dependency. It is already a transitive dependency of the LSP package, so this is not actually adding anything, but our linters enforce the good practice of declaring explicit dependencies
2. Added the run handler when the feature flag is enabled. It will ask the server for commands and execute them using JSON RPC to receive notifications back about the state of tests, which it then uses to update the UI

### Automated Tests

The easiest way I found of testing this is to create a small fake test server script, so that we can mock the resolve commands return and still have something that will print JSON RPC notifications.

I also extract a common part of the test.