Use map compact instead of filter_map in activation script (#3321)
### Motivation

The activation script is failing when no Shadowenv environment is activated because it then falls back to system Ruby (2.6) and `filter_map` doesn't exist in it.

There are better long term solutions to this, like trying to fallback to chruby when Shadowenv fails or offering other options, but for now I would really like to remove these errors from telemetry since we finally managed to fix the encoding related problems too.

### Implementation

Switched to map/compact instead of `filter_map`.