Avoiding rejecting promise if test command failed (#3334)
### Motivation

When we execute a test that fails, the command will exit with a code different than zero. However, that's expected, since the tests failed.

We shouldn't reject the promise if the tests failed (only if some of the test controller's internals did). Otherwise, the explorer won't finalize the run properly.

### Implementation

We just need to resolve every time regardless of what the exit code was.