Automatically discover children for test files that weren't expanded (#3336)
### Motivation

I noticed that executing entire test directories through the explorer wasn't working. This is due to #3335 and the aspect that we were not automatically discovering the children of test files if they were never expanded.

That meant that, we had no knowledge of their children existing, but when running the tests we are reporting results for those children we know nothing about.

We need to automatically resolve children for test files that haven't been expanded otherwise the items we need to update don't exist.

### Implementation

It's just a matter of invoking our resolve handler if there are no children in the test file item.

### Automated Tests

Added a test.