Handle changing visibility post method definition (#3339)
### Motivation

I noticed we didn't handle changing method visibility post definition. This PR adds support for it.

### Implementation

The idea is to check if there are string or symbol arguments being passed and then change the visibility of the methods passed.

Otherwise, we continue doing the same as before.

### Automated Tests

Added a test.