Allow disabling the Sorbet detection via config
This change introduces a new initialization option for disabling the
automatic detection of a type checker. This behavior was initially
implemented through an environment variable (which this change maintains
for backwards-compatibility), but that isn't the easiest thing to
configure.

By making it a part of initialization, we can now declaratively set the
flag with all other configuration for the server.