Add minitest/spec code lens test command resolution for minitest/spec
Changing the ids of the test items of minitest spec to what actually
that DSL generates in the minitest methods, doing that we reuse the
command test resolution of minitest test style. Also change the command
test resolution to be based on id instead of labels to have a nice tree
with names in the test pannel, but having enough metadata to parse the
spec style tests in the test resolution.