Add test add-on command resolution docs (#3616)
### Motivation

Step towards #3421

This PR adds the second part of the test add-on documentation, which is command resolution. I tried to explain in a generic way, showing some examples of what command resolution is supposed to do.

Please let me know if this is not clear enough and if the examples need to be enhanced.