Account for module wrapping Minitest specs (#3715)
### Motivation

Closes #3712

When we find no test classes, we were considering the describe groups to start at index 1 in our stack of test groups. If you have a module wrapping those specs, that won't be true as there will be a corresponding `nil` entry in the stack (which we use to balance the stack).

We need to search for what is actually the first index to account for this possibility.

### Implementation

Started searching for what is the first spec group, rather than assuming that it's always the first one. I also moved our `in_spec_context?` check since we want to allow the `describe` to be discovered even if the nesting is not empty and we haven't yet found a spec group (which is what happens when there's only a top level module).

### Automated Tests

Added a test that reproduces the scenario.