Speed up checking whether entry is coming from a dependency
Doing filtering brefore fuzzy matching didn't improve performance,
because retrieving full paths of entries is slow.

Since the file paths are escaped in URIs, they're currently being
unescaped with Windows handling in order to retrieve the raw path. This
is extra work, because we know the raw path at the time we're building
the URI, we just didn't store that information.

Storing the raw path in the URI and retrieving it speeds up workspace
symbol search ~4x in my application.