Implement workspace/didChangeConfiguration and workspace/configuration
This adds support for the pull model of dynamic configuration changes,
as described here: https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workspace_configuration

Upon receiving `workspace/didChangeConfiguration` message the Ruby LSP
server sends `workspace/configuration` to the client. The client
replies. This reply is different than other replies handled by Ruby LSP
so far, as it does not include `method` and needs to be associated with
the request via `id`. This required implementing a collection of
server-sent requests, to be able to match.

The `result` of the reply is not a hash, instead it's an array of
hashes, so it needed to be handled separately as well.

Technically, upon receiving `workspace/configuration` the server should
check if it should register or unregister some capabilities. I intended
to do that too, but it started to become messy and also I did not have a
way to properly test it.