Ensure binmode and sync on LSP reporter socket (#3820)
### Motivation

I believe this will close #3760

Since we're using JSON RPC to communicate test results between server and extension, we have to turn on

- binmode: so that Windows preserves the `\r\n\r\n` sequences that are the key request separators
- sync: so that the content of the pipe is not buffered, but immediately flushed

We already do this one very stdio we use and we forgot to do the same for the socket used for tests.

### Implementation

Added `binmode` and `sync` to our LSP reporter socket.

### Automated Tests

I believe that turning `binmode` on our fake test client would've been enough to catch this.