Avoid failing if the git extension is disabled (#3822)
### Motivation

Closes #3685

If the git extension is disabled, trying to invoke `getAPI` may throw an error like the one reported in the issue.

### Implementation

We can just wrap that call in a try/catch block since we fall back to the default, which is to consider the git root to be in the same place as the workspace path.